refactor(go): Extract util file
