+++ lib/datasource/go/index.ts
@@ -1,153 +1,14 @@
-import URL from 'url';
-import { PlatformId } from '../../constants';
 import { logger } from '../../logger';
-import * as hostRules from '../../util/host-rules';
 import { regEx } from '../../util/regex';
-import { trimTrailingSlash } from '../../util/url';
-import { BitBucketTagsDatasource } from '../bitbucket-tags';
 import * as github from '../github-tags';
 import * as gitlab from '../gitlab-tags';
 import type { DigestConfig, GetReleasesConfig, ReleaseResult } from '../types';
-import { http } from './common';
 import * as goproxy from './goproxy';
-import type { DataSource } from './types';
+import { bitbucket, getDatasource } from './util';
 
 export { id } from './common';
 
 export const customRegistrySupport = false;
-
-const gitlabHttpsRegExp = regEx(
-  /^(?<httpsRegExpUrl>https:\/\/[^/]*gitlab\.[^/]*)\/(?<httpsRegExpName>.+?)[/]?$/
-);
-const gitlabRegExp = regEx(
-  /^(?<regExpUrl>gitlab\.[^/]*)\/(?<regExpPath>.+?)[/]?$/
-);
-const bitbucket = new BitBucketTagsDatasource();
-
-async function getDatasource(goModule: string): Promise<DataSource | null> {
-  if (goModule.startsWith('gopkg.in/')) {
-    const [pkg] = goModule.replace('gopkg.in/', '').split('.');
-    if (pkg.includes('/')) {
-      return { datasource: github.id, lookupName: pkg };
-    }
-    return {
-      datasource: github.id,
-      lookupName: `go-${pkg}/${pkg}`,
-    };
-  }
-  if (goModule.startsWith('github.com/')) {
-    const split = goModule.split('/');
-    const lookupName = split[1] + '/' + split[2];
-    return {
-      datasource: github.id,
-      lookupName,
-    };
-  }
-
-  if (goModule.startsWith('bitbucket.org/')) {
-    const split = goModule.split('/');
-    const lookupName = split[1] + '/' + split[2];
-    return {
-      datasource: bitbucket.id,
-      lookupName,
-    };
-  }
-
-  const pkgUrl = `https://${goModule}?go-get=1`;
-  const res = (await http.get(pkgUrl)).body;
-  const sourceMatch = regEx(
-    `<meta\\s+name="go-source"\\s+content="([^\\s]+)\\s+([^\\s]+)`
-  ).exec(res);
-  if (sourceMatch) {
-    const [, prefix, goSourceUrl] = sourceMatch;
-    if (!goModule.startsWith(prefix)) {
-      logger.trace({ goModule }, 'go-source header prefix not match');
-      return null;
-    }
-    logger.debug({ goModule, goSourceUrl }, 'Go lookup source url');
-    if (goSourceUrl?.startsWith('https://github.com/')) {
-      return {
-        datasource: github.id,
-        lookupName: goSourceUrl
-          .replace('https://github.com/', '')
-          .replace(regEx(/\/$/), ''),
-      };
-    }
-    const gitlabUrl =
-      gitlabHttpsRegExp.exec(goSourceUrl)?.groups?.httpsRegExpUrl;
-    const gitlabUrlName =
-      gitlabHttpsRegExp.exec(goSourceUrl)?.groups?.httpsRegExpName;
-    const gitlabModuleName = gitlabRegExp.exec(goModule)?.groups?.regExpPath;
-
-    if (gitlabUrl && gitlabUrlName) {
-      if (gitlabModuleName?.startsWith(gitlabUrlName)) {
-        return {
-          datasource: gitlab.id,
-          registryUrl: gitlabUrl,
-          lookupName: gitlabModuleName,
-        };
-      }
-      return {
-        datasource: gitlab.id,
-        registryUrl: gitlabUrl,
-        lookupName: gitlabUrlName,
-      };
-    }
-
-    const opts = hostRules.find({
-      hostType: PlatformId.Gitlab,
-      url: goSourceUrl,
-    });
-    if (opts.token) {
-      // get server base url from import url
-      const parsedUrl = URL.parse(goSourceUrl);
-
-      // split the go module from the URL: host/go/module -> go/module
-      const split = goModule.split('/');
-      const lookupName = split[1] + '/' + split[2];
-
-      const registryUrl = `${parsedUrl.protocol}//${parsedUrl.host}`;
-
-      return {
-        datasource: gitlab.id,
-        registryUrl,
-        lookupName,
-      };
-    }
-  } else {
-    // GitHub Enterprise only returns a go-import meta
-    const importMatch = regEx(
-      `<meta\\s+name="go-import"\\s+content="([^\\s]+)\\s+([^\\s]+)\\s+([^\\s]+)">`
-    ).exec(res);
-    if (importMatch) {
-      const [, prefix, , goImportURL] = importMatch;
-      if (!goModule.startsWith(prefix)) {
-        logger.trace({ goModule }, 'go-import header prefix not match');
-        return null;
-      }
-      logger.debug({ goModule, goImportURL }, 'Go lookup import url');
-
-      // get server base url from import url
-      const parsedUrl = URL.parse(goImportURL);
-
-      // split the go module from the URL: host/go/module -> go/module
-      const lookupName = trimTrailingSlash(parsedUrl.pathname)
-        .replace(regEx(/\.git$/), '')
-        .split('/')
-        .slice(-2)
-        .join('/');
-
-      return {
-        datasource: github.id,
-        registryUrl: `${parsedUrl.protocol}//${parsedUrl.host}`,
-        lookupName,
-      };
-    }
-
-    logger.trace({ goModule }, 'No go-source or go-import header found');
-  }
-  return null;
-}
 
 /**
  * go.getReleases
